+++ lib/QECCore/src/QECCore.jl
@@ -4,7 +4,7 @@ using SparseArrays
 using LinearAlgebra
 using Combinatorics
 using Graphs
-using Random: GLOBAL_RNG, AbstractRNG, randperm, rand
+using Random: GLOBAL_RNG, AbstractRNG, randperm, rand, MersenneTwister, randperm
 
 using DocStringExtensions
 
@@ -21,7 +21,7 @@ DelfosseReichardt, DelfosseReichardtRepCode, DelfosseReichardt823, QuantumTanner
 TillichZemor, random_TillichZemor_code, CirculantBivariateBicycle
 
 # Classical Codes
-export RepCode, ReedMuller, RecursiveReedMuller, Golay, Hamming
+export RepCode, ReedMuller, RecursiveReedMuller, Golay, Hamming, GallagerLDPC
 
 # utilities
 export search_self_orthogonal_rm_codes
@@ -31,8 +31,9 @@ include("codes/util.jl")
 
 # Classical Codes
 include("codes/classical/hamming.jl")
-include("codes/classical/reptetion.jl")
+include("codes/classical/repetition.jl")
 include("codes/classical/golay.jl")
+include("codes/classical/gallager.jl")
 
 # Quantum Codes
 include("codes/quantum/css.jl")